use external tracing-ffi-schema
